% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{ldnotice}
\alias{ldnotice}
\title{Get a Notice}
\usage{
ldnotice(notice, ...)
}
\arguments{
\item{notice}{A single numeric value specifying a notice ID.}

\item{...}{Additional arguments passed to \code{\link{lumenHTTP}}.}
}
\value{
A list of class \dQuote{lumen_notice}. A \code{summary} method will display some essential details of the notice.
}
\description{
Get a Lumen Database Notice by ID
}
\details{
This function retrieves the details of a Lumen Database notice based upon its Notice ID number. To search for notices, use \code{\link{ldsearch}} instead.
}
\examples{
\dontrun{
n <- ldnotice(1)
summary(n)
}
}

