% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http.R
\name{ldtopics}
\alias{ldtopics}
\title{Get Topics}
\usage{
ldtopics(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{lumenHTTP}}.}
}
\value{
A list of objects of class \dQuote{lumen_topic}. The default \code{print} method will display some essential details of each topic.
}
\description{
Get a list of Lumen Database Topics
}
\details{
This function retrieves a list of \dQuote{topics} used by the Lumen Database.
}
\examples{
\dontrun{
x <- ldtopics()
x
}
}

